HIRE-96: made initialization of database configurable

+++ src/main/java/conf/DatastoreProvider.java
@@ -1,9 +1,8 @@
 package conf;
 
+import com.github.fakemongo.Fongo;
 import com.google.inject.Provider;
 import com.google.inject.Singleton;
-
-import com.github.fakemongo.Fongo;
 import com.mongodb.MongoClient;
 import com.mongodb.MongoCredential;
 import com.mongodb.ServerAddress;
@@ -30,6 +29,8 @@ public class DatastoreProvider implements Provider<Datastore> {
   private static final String MONGODB_DBNAME = "ninja.mongodb.dbname";
   private static final String MONGODB_AUTHDB = "ninja.mongodb.authdb";
   private static final String MONGODB_USE_IN_MEMORY = "ninja.mongodb.useInMemory";
+  private static final String MONGODB_INITIALIZE_DB = "ninja.mongodb.initializedb";
+
   private NinjaProperties properties;
   private MongoClient mongoClient;
   private Morphia morphia;
@@ -63,7 +64,9 @@ public class DatastoreProvider implements Provider<Datastore> {
         MongoCredential.createCredential(username, authdb, password.toCharArray());
     mongoClient = new MongoClient(new ServerAddress(host, port), Arrays.asList(credentials));
     databaseName = properties.get(MONGODB_DBNAME);
-    mongoClient.dropDatabase(databaseName);
+    if (properties.getBooleanWithDefault(MONGODB_INITIALIZE_DB, false)) {
+      mongoClient.dropDatabase(databaseName);
+    }
   }
 
   public Datastore get() {
